Issue 18 - refactor mount list file IO to use Mesos checkpoint code/template:
After comparing the output of the current file IO process to the checkpoint
method, the output is identical and behaves as expected. Tested multiple
scenarios in init and recovery.

Removed temporary exploratory code that compares the current json file
and replaced with the checkpoint implementation.

Please note that code has been copied from the internal mesos namespace. This
code will be deleted once the checkpoint process is exposed publically.
This code has been annotated with comments in code.